[imp] separate_page_role_user PSG-22

+++ views/categories/category.view.php
@@ -1,8 +1,10 @@
 <div class="card-header pl-5 pr-5 py-3 d-flex justify-content-between">
     <h4 class="font-weight-bold text-primary">Category</h4>
-    <button type="button" class="btn btn-outline-primary" data-toggle="modal" data-target="#medalCreate">
-        <i class="bi bi-plus-circle mr-2"></i>Add new Category
-    </button>
+    <?php if ($_SESSION['user']['role'] != 'cashier') : ?>
+        <button type="button" class="btn btn-outline-primary" data-toggle="modal" data-target="#medalCreate">
+            <i class="bi bi-plus-circle mr-2"></i>Add new Category
+        </button>
+    <?php endif; ?>
 </div>
 
 
@@ -24,36 +26,37 @@
             ?>
 
                 <!-- popup update  -->
+                <?php if ($_SESSION['user']['role'] != 'cashier') : ?>
+                    <div class="modal fade" id="exampleModal<?= $cate['cate_id'] ?>" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
+                        <div class="modal-dialog" role="document">
+                            <div class="modal-content">
+                                <div class="modal-header">
+                                    <h5 class="modal-title text-danger ml-3" id="exampleModalLabel">Edit Categry</h5>
+                                    <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                                        <span aria-hidden="true">&times;</span>
+                                    </button>
+                                </div>
 
-                <div class="modal fade" id="exampleModal<?= $cate['cate_id'] ?>" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
-                    <div class="modal-dialog" role="document">
-                        <div class="modal-content">
-                            <div class="modal-header">
-                                <h5 class="modal-title text-danger ml-3" id="exampleModalLabel">Edit Categry</h5>
-                                <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-                                    <span aria-hidden="true">&times;</span>
-                                </button>
-                            </div>
-
-                            <div class="m-5 d-flex justify-content-center flex-column">
-                                <form action="controllers/categories/update_category.controller.php" class="d-flex flex-xl-column" method="post">
-                                    <input type="hidden" name="id" value="<?= $cate['cate_id'] ?>">
-                                    <div class="form-group">
-                                        <label for="recipient-name" class="col-form-label">Name</label>
-                                        <input type="text" class="form-control" value="<?= $cate['cate_name'] ?>" name="category" placeholder="Category_name">
-                                    </div>
-                                    <div class="form-group">
-                                        <label for="message-text" class="col-form-label">Description</label>
-                                        <input type="text" class="form-control" value="<?= $cate['cate_desc'] ?>" name="description" placeholder="Category_name">
-                                    </div>
-                                    <div class="modal-footer">
-                                        <button type="submit" class="btn btn-primary">Save changes</button>
-                                    </div>
-                                </form>
+                                <div class="m-5 d-flex justify-content-center flex-column">
+                                    <form action="controllers/categories/update_category.controller.php" class="d-flex flex-xl-column" method="post">
+                                        <input type="hidden" name="id" value="<?= $cate['cate_id'] ?>">
+                                        <div class="form-group">
+                                            <label for="recipient-name" class="col-form-label">Name</label>
+                                            <input type="text" class="form-control" value="<?= $cate['cate_name'] ?>" name="category" placeholder="Category_name">
+                                        </div>
+                                        <div class="form-group">
+                                            <label for="message-text" class="col-form-label">Description</label>
+                                            <input type="text" class="form-control" value="<?= $cate['cate_desc'] ?>" name="description" placeholder="Category_name">
+                                        </div>
+                                        <div class="modal-footer">
+                                            <button type="submit" class="btn btn-primary">Save changes</button>
+                                        </div>
+                                    </form>
+                                </div>
                             </div>
                         </div>
                     </div>
-                </div>
+                <?php endif; ?>
 
                 <!-- popup detail  -->
 
@@ -104,11 +107,13 @@
                         <div class="btn-group">
                             <div class="btn-group">
 
-                                <i class="bi bi-pencil-square text-success btn btn-lg ml-1" data-toggle="modal" data-target="#exampleModal<?= $cate['cate_id'] ?>"></i>
-
-                                <a href="../../controllers/categories/delete_category.controller.php?id=<?= $cate['cate_id'] ?> " onclick="return confirm('Do you want to delete this product?')">
-                                    <i class="bi bi-trash3 text-danger btn btn-lg ml-1"></i></a>
-
+                                <?php if ($_SESSION['user']['role'] != 'cashier') : ?>
+                                    <i class="bi bi-pencil-square text-success btn btn-lg ml-1" data-toggle="modal" data-target="#exampleModal<?= $cate['cate_id'] ?>"></i>
+                                <?php endif; ?>
+                                <?php if ($_SESSION['user']['role'] != 'cashier') : ?>
+                                    <a href="../../controllers/categories/delete_category.controller.php?id=<?= $cate['cate_id'] ?> " onclick="return confirm('Do you want to delete this product?')">
+                                        <i class="bi bi-trash3 text-danger btn btn-lg ml-1"></i></a>
+                                <?php endif; ?>
                                 <i class="bi bi-eye-fill text-info btn btn-lg ml-1" data-toggle="modal" data-target="#exampleModals<?= $cate['cate_id'] ?>"></i>
 
                             </div>
@@ -118,45 +123,38 @@
         </tbody>
     </table>
 </div>
-</div>
-</div>
-
-</div>
-
 
 
 <!-- popup create  -->
+<?php if ($_SESSION['user']['role'] != 'cashier') : ?>
+    <div class="modal fade" id="medalCreate" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
+        <div class="modal-dialog">
+            <div class="modal-content">
+                <div class="modal-header">
+                    <h5 class="modal-title text-danger" id="exampleModalLabel">Add Category</h5>
+                    <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                        <span aria-hidden="true">&times;</span>
+                    </button>
+                </div>
+                <form action="controllers/categories/create_category.controller.php" class="d-flex flex-xl-column" method="post">
+                    <div class="modal-body">
 
-<div class="modal fade" id="medalCreate" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
-    <div class="modal-dialog">
-        <div class="modal-content">
-            <div class="modal-header">
-                <h5 class="modal-title text-danger" id="exampleModalLabel">Add Category</h5>
-                <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-                    <span aria-hidden="true">&times;</span>
-                </button>
+                        <div class="form-group">
+                            <label for="title" class="text-primary m-2">Name</label>
+                            <input type="text" class="form-control" name="category" placeholder="Cate name">
+                        </div>
+                        <div class="form-group">
+                            <label for="comment" class="text-primary m-2">Comment</label>
+                            <input type="text" class="form-control" name="description" placeholder="comment Cate ">
+
+                        </div>
+                    </div>
+                    <div class="modal-footer">
+                        <button type="submit" class="btn btn-primary">Submit</button>
+                    </div>
+                </form>
             </div>
-            <form action="controllers/categories/create_category.controller.php" class="d-flex flex-xl-column" method="post">
-                <div class="modal-body">
-
-                    <div class="form-group">
-                        <label for="title" class="text-primary m-2">Name</label>
-                        <input type="text" class="form-control" name="category" placeholder="Cate name">
-                    </div>
-                    <div class="form-group">
-                        <label for="comment" class="text-primary m-2">Comment</label>
-                        <input type="text" class="form-control" name="description" placeholder="comment Cate ">
-
-                    </div>
-                </div>
-                <div class="modal-footer">
-                    <button type="submit" class="btn btn-primary">Submit</button>
-                </div>
-            </form>
         </div>
+
     </div>
-
-</div>
-</div>
-
-</div>
+<?php endif; ?>
